requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.3
asttokens==2.4.1
astunparse==1.6.3
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==3.3.2
clarifai==11.2.3
clarifai-grpc==11.3.2
clarifai-protocol==0.0.22
click==8.1.7
comm==0.2.1
contextlib2==21.6.0
contourpy==1.2.0
cssselect2==0.7.0
cvxopt==1.3.2
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dm-tree==0.1.8
docopt==0.6.2
docstring_parser==0.16
etils==1.9.2
executing==2.0.1
fastjsonschema==2.21.1
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==24.3.25
fonttools==4.47.2
frozenlist==1.5.0
fsspec==2024.6.1
gast==0.5.4
gobject==0.1.0
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.8
httpx==0.28.1
idna==3.6
imageio==2.34.0
immutabledict==4.2.0
importlib_resources==6.4.0
ipykernel==6.29.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
keras==3.1.1
kiwisolver==1.4.5
lazy_loader==0.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.12
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.1.3
ml-dtypes==0.3.2
multidict==6.4.3
namex==0.0.7
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
openai==1.74.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pillow_heif==0.22.0
pipreqs==0.5.0
platformdirs==4.1.0
process-image==0.0.3
promise==2.3
prompt-toolkit==3.0.43
propcache==0.3.1
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pycparser==2.21
pydantic==2.11.3
pydantic_core==2.33.1
pydyf==0.8.0
Pygments==2.17.2
pyparsing==3.1.1
pyphen==0.14.0
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2023.4
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.36.2
requests==2.32.3
rich==13.7.1
rpds-py==0.24.0
schema==0.7.5
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.13.1
setuptools==69.2.0
simple_parsing==0.1.5
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-datasets==4.9.6
tensorflow-metadata==1.15.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.2.12
tinycss2==1.2.1
toml==0.10.2
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2023.4
urllib3==2.2.1
wcwidth==0.2.13
weasyprint==61.0
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.43.0
wrapt==1.16.0
yarg==0.1.9
yarl==1.19.0
zipp==3.19.2
zopfli==0.2.3
